Platform/Sgi: parameterize the iort table
In preparation of adding multiple SMMU, IORT and PCIe root complex nodes
into the IORT ACPI table, parameterize the existing IORT table. The SMMU
interrupt numbers, device ID and base addresses are all parameterized
using helper macros. PCDs for these parameters are defined and the
platforms can define the value of these PCDs.
While we are at it fix the Coherency flags and Number of Ids fields of
various nodes.

Signed-off-by: Vivek Gautam <vivek.gautam@arm.com>
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: If631ad3695fb5c7a87c11906fa2331d328fd3495