Platform/Sgi: generate pcie rootbridge topology layout dynamically
Add support to generate RootBridge layout dynamically from the topology
information passed to UEFI. Use a template to generate the basic layout
of the topology and then use the topology information to update
PcieHostBridgeLib specific datastructure.

This file is share by all the platforms on Sgi platforms, add a Pcd flag
to enable or disable this feature for the plarforms which does not use
this feature.

Signed-off-by: Nishant Sharma <nishant.sharma@arm.com>
Change-Id: I39d64bf56048e3509f379c6e22514fe12898b2c3