Platform/Sgi: Fix cache ID macro's comments and params
For computing cache ID, the package ID, clusted ID and cpu IDs are used.
For a private cache, cluster and cpu ID could be zero or non-zero.
cluster or cpu ID being zero indicates the first (as cluster or cpu
numbering starts from 0 to n-1, where n is the count of clusters or
cpus) block of cluster or cpu respectively. Fix the comments stating
that cpu and cluster ID being zero indicates a shared cache. Instead,
'-1' is to be passed for cluster and cpu ID for shared caches.

Additionally, update the macro's CoreId parameter to CpuId to match the
description in the comments.

Signed-off-by: Rohit Mathew <rohit.mathew@arm.com>
Change-Id: I935ef51e06a3696da171f2775e2d03e5562913a5