Platform/Sgi: add CXL specific ACPI tables
PcieAcpiTableGenerator, creates SSDT table for PCIe, based on data at
runtime. For CXL enabled platform, there has to be presence of CXL
specific ACPI objects.

ACPI0017 is a SW entity under System Bus of ACPI tree, that indicates
the presence of CEDT table.

ACPI0016 is a SW entity under System Bus of ACPI tree, that represents
a CXL Host Bridge.

The _UID object under a CXL Host Bridge object, when evaluated, must
match the UID field in the associated CHBS entry. In this patch,
the UID across the objects, is Zero.

Signed-off-by: Sayanta Pattanayak <sayanta.pattanayak@arm.com>
Change-Id: Ic90785e6b453584a642e86ddfb3a4bb3cc6f0a83