Platform/Sgi: Add initial support for RD-Fremont-Cfg1 platform
Arm's RD-Fremont-Cfg1 platform is a variant of the RD-Fremont platform.
Compared to RD-Fremont platform, RD-Fremont-Cfg1 has a reduced core
count of eight Neoverse Poseidon CPUs and a smaller interconnect mesh.
As part of the initial platform support for RD-Fremont-Cfg1 platform,
add the corresponding ACPI tables, platform and flash description files.

Signed-off-by: Shriram K <shriram.k@arm.com>
Signed-off-by: Vivek Gautam <vivek.gautam@arm.com>
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: I2f00f28e38927cc15cb27e8f8e2f3580ceb696b6