Platform/Sgi: Enable auto generation and installation of MCFG table.
With the dynamic configuration of PCIe controller ECAM and MMIO regions.
MCFG cannot be compiled statically. To resolve this issue the MCFG table
should be generated and installed dynamically from the data provided by
the previous boot components.

This patch add the support to parse the dynamically generated PCIe
region and bus info and add entries to the MCFG table for each segment
available on the platform.

Signed-off-by: Nishant Sharma <nishant.sharma@arm.com>
Change-Id: Ice9a2aada301994a6b1deb1d9ba3fc38ffab24fa